+++ main.go
@@ -20,7 +20,7 @@ import (
 	"github.com/spf13/viper"
 )
 
-const atlantisVersion = "0.6.0"
+const atlantisVersion = "0.7.0"
 
 func main() {
 	v := viper.New()