+++ server/events/project_command_runner.go
@@ -55,6 +55,14 @@ type StepRunner interface {
 	Run(ctx models.ProjectCommandContext, extraArgs []string, path string) (string, error)
 }
 
+//go:generate pegomock generate -m --use-experimental-model-gen --package mocks -o mocks/mock_custom_step_runner.go CustomStepRunner
+
+// CustomStepRunner runs custom run steps.
+type CustomStepRunner interface {
+	// Run cmd in path.
+	Run(ctx models.ProjectCommandContext, cmd string, path string) (string, error)
+}
+
 //go:generate pegomock generate -m --use-experimental-model-gen --package mocks -o mocks/mock_webhooks_sender.go WebhooksSender
 
 // WebhooksSender sends webhook.
@@ -76,18 +84,16 @@ type ProjectCommandRunner interface {
 
 // DefaultProjectCommandRunner implements ProjectCommandRunner.
 type DefaultProjectCommandRunner struct {
-	Locker                   ProjectLocker
-	LockURLGenerator         LockURLGenerator
-	InitStepRunner           StepRunner
-	PlanStepRunner           StepRunner
-	ApplyStepRunner          StepRunner
-	RunStepRunner            StepRunner
-	PullApprovedChecker      runtime.PullApprovedChecker
-	WorkingDir               WorkingDir
-	Webhooks                 WebhooksSender
-	WorkingDirLocker         WorkingDirLocker
-	RequireApprovalOverride  bool
-	RequireMergeableOverride bool
+	Locker              ProjectLocker
+	LockURLGenerator    LockURLGenerator
+	InitStepRunner      StepRunner
+	PlanStepRunner      StepRunner
+	ApplyStepRunner     StepRunner
+	RunStepRunner       CustomStepRunner
+	PullApprovedChecker runtime.PullApprovedChecker
+	WorkingDir          WorkingDir
+	Webhooks            WebhooksSender
+	WorkingDirLocker    WorkingDirLocker
 }
 
 // Plan runs terraform plan for the project described by ctx.
@@ -100,7 +106,7 @@ func (p *DefaultProjectCommandRunner) Plan(ctx models.ProjectCommandContext) mod
 		Failure:     failure,
 		RepoRelDir:  ctx.RepoRelDir,
 		Workspace:   ctx.Workspace,
-		ProjectName: ctx.GetProjectName(),
+		ProjectName: ctx.ProjectName,
 	}
 }
 
@@ -114,7 +120,7 @@ func (p *DefaultProjectCommandRunner) Apply(ctx models.ProjectCommandContext) mo
 		ApplySuccess: applyOut,
 		RepoRelDir:   ctx.RepoRelDir,
 		Workspace:    ctx.Workspace,
-		ProjectName:  ctx.GetProjectName(),
+		ProjectName:  ctx.ProjectName,
 	}
 }
 
@@ -149,17 +155,7 @@ func (p *DefaultProjectCommandRunner) doPlan(ctx models.ProjectCommandContext) (
 		return nil, "", DirNotExistErr{RepoRelDir: ctx.RepoRelDir}
 	}
 
-	// Use default stage unless another workflow is defined in config
-	stage := p.defaultPlanStage()
-	if ctx.ProjectConfig != nil && ctx.ProjectConfig.Workflow != nil {
-		ctx.Log.Debug("project configured to use workflow %q", *ctx.ProjectConfig.Workflow)
-		configuredStage := ctx.GlobalConfig.GetPlanStage(*ctx.ProjectConfig.Workflow)
-		if configuredStage != nil {
-			ctx.Log.Debug("project will use the configured stage for that workflow")
-			stage = *configuredStage
-		}
-	}
-	outputs, err := p.runSteps(stage.Steps, ctx, projAbsPath)
+	outputs, err := p.runSteps(ctx.Steps, ctx, projAbsPath)
 	if err != nil {
 		if unlockErr := lockAttempt.UnlockFn(); unlockErr != nil {
 			ctx.Log.Err("error unlocking state after plan error: %v", unlockErr)
@@ -214,21 +210,7 @@ func (p *DefaultProjectCommandRunner) doApply(ctx models.ProjectCommandContext) 
 		return "", "", DirNotExistErr{RepoRelDir: ctx.RepoRelDir}
 	}
 
-	// Figure out what our apply requirements are.
-	var applyRequirements []string
-	if p.RequireApprovalOverride || p.RequireMergeableOverride {
-		// If any server flags are set, they override project config.
-		if p.RequireMergeableOverride {
-			applyRequirements = append(applyRequirements, raw.MergeableApplyRequirement)
-		}
-		if p.RequireApprovalOverride {
-			applyRequirements = append(applyRequirements, raw.ApprovedApplyRequirement)
-		}
-	} else if ctx.ProjectConfig != nil {
-		// Else we use the project config if it's set.
-		applyRequirements = ctx.ProjectConfig.ApplyRequirements
-	}
-	for _, req := range applyRequirements {
+	for _, req := range ctx.ApplyRequirements {
 		switch req {
 		case raw.ApprovedApplyRequirement:
 			approved, err := p.PullApprovedChecker.PullIsApproved(ctx.BaseRepo, ctx.Pull) // nolint: vetshadow
@@ -251,15 +233,7 @@ func (p *DefaultProjectCommandRunner) doApply(ctx models.ProjectCommandContext) 
 	}
 	defer unlockFn()
 
-	// Use default stage unless another workflow is defined in config
-	stage := p.defaultApplyStage()
-	if ctx.ProjectConfig != nil && ctx.ProjectConfig.Workflow != nil {
-		configuredStage := ctx.GlobalConfig.GetApplyStage(*ctx.ProjectConfig.Workflow)
-		if configuredStage != nil {
-			stage = *configuredStage
-		}
-	}
-	outputs, err := p.runSteps(stage.Steps, ctx, absPath)
+	outputs, err := p.runSteps(ctx.Steps, ctx, absPath)
 	p.Webhooks.Send(ctx.Log, webhooks.ApplyResult{ // nolint: errcheck
 		Workspace: ctx.Workspace,
 		User:      ctx.User,
@@ -271,27 +245,4 @@ func (p *DefaultProjectCommandRunner) doApply(ctx models.ProjectCommandContext) 
 		return "", "", fmt.Errorf("%s\n%s", err, strings.Join(outputs, "\n"))
 	}
 	return strings.Join(outputs, "\n"), "", nil
-}
-
-func (p DefaultProjectCommandRunner) defaultPlanStage() valid.Stage {
-	return valid.Stage{
-		Steps: []valid.Step{
-			{
-				StepName: "init",
-			},
-			{
-				StepName: "plan",
-			},
-		},
-	}
-}
-
-func (p DefaultProjectCommandRunner) defaultApplyStage() valid.Stage {
-	return valid.Stage{
-		Steps: []valid.Step{
-			{
-				StepName: "apply",
-			},
-		},
-	}
 }