+++ server/events/vcs/bitbucketserver/client.go
@@ -15,7 +15,7 @@ import (
 
 	"github.com/pkg/errors"
 	"github.com/runatlantis/atlantis/server/events/models"
-	"gopkg.in/go-playground/validator.v9"
+	validator "gopkg.in/go-playground/validator.v9"
 )
 
 // maxCommentLength is the maximum number of chars allowed by Bitbucket in a