+++ server/events_controller_e2e_test.go
@@ -3,8 +3,6 @@ package server_test
 import (
 	"bytes"
 	"fmt"
-	"github.com/hashicorp/go-getter"
-	"github.com/runatlantis/atlantis/server/events/db"
 	"io/ioutil"
 	"net/http"
 	"net/http/httptest"
@@ -14,6 +12,10 @@ import (
 	"regexp"
 	"strings"
 	"testing"
+
+	getter "github.com/hashicorp/go-getter"
+	"github.com/runatlantis/atlantis/server/events/db"
+	"github.com/runatlantis/atlantis/server/events/yaml/valid"
 
 	"github.com/google/go-github/github"
 	. "github.com/petergtz/pegomock"
@@ -291,12 +293,29 @@ func TestGitHubWorkflow(t *testing.T) {
 				"exp-output-merge.txt",
 			},
 		},
+		{
+			Description:   "server-side cfg",
+			RepoDir:       "server-side-cfg",
+			ExpAutomerge:  false,
+			ExpAutoplan:   true,
+			ModifiedFiles: []string{"main.tf"},
+			Comments: []string{
+				"atlantis apply -w staging",
+				"atlantis apply -w default",
+			},
+			ExpReplies: []string{
+				"exp-output-autoplan.txt",
+				"exp-output-apply-staging-workspace.txt",
+				"exp-output-apply-default-workspace.txt",
+				"exp-output-merge.txt",
+			},
+		},
 	}
 	for _, c := range cases {
 		t.Run(c.Description, func(t *testing.T) {
 			RegisterMockTestingT(t)
 
-			ctrl, vcsClient, githubGetter, atlantisWorkspace := setupE2E(t)
+			ctrl, vcsClient, githubGetter, atlantisWorkspace := setupE2E(t, c.RepoDir)
 			// Set the repo to be cloned through the testing backdoor.
 			repoDir, headSHA, cleanup := initializeRepo(t, c.RepoDir)
 			defer cleanup()
@@ -355,7 +374,7 @@ func TestGitHubWorkflow(t *testing.T) {
 	}
 }
 
-func setupE2E(t *testing.T) (server.EventsController, *vcsmocks.MockClient, *mocks.MockGithubPullGetter, *events.FileWorkspace) {
+func setupE2E(t *testing.T, repoDir string) (server.EventsController, *vcsmocks.MockClient, *mocks.MockGithubPullGetter, *events.FileWorkspace) {
 	allowForkPRs := false
 	dataDir, cleanup := TempDir(t)
 	defer cleanup()
@@ -391,8 +410,15 @@ func setupE2E(t *testing.T) (server.EventsController, *vcsmocks.MockClient, *moc
 		TestingOverrideHeadCloneURL: "override-me",
 	}
 
-	defaultTFVersion := terraformClient.Version()
+	defaultTFVersion := terraformClient.DefaultVersion()
 	locker := events.NewDefaultWorkingDirLocker()
+	parser := &yaml.ParserValidator{}
+	globalCfg := valid.NewGlobalCfg(true, false, false)
+	expCfgPath := filepath.Join(absRepoPath(t, repoDir), "repos.yaml")
+	if _, err := os.Stat(expCfgPath); err == nil {
+		globalCfg, err = parser.ParseGlobalCfg(expCfgPath, globalCfg)
+		Ok(t, err)
+	}
 	commandRunner := &events.DefaultCommandRunner{
 		ProjectCommandRunner: &events.DefaultProjectCommandRunner{
 			Locker:           projectLocker,
@@ -426,15 +452,14 @@ func setupE2E(t *testing.T) (server.EventsController, *vcsmocks.MockClient, *moc
 		AllowForkPRs:             allowForkPRs,
 		AllowForkPRsFlag:         "allow-fork-prs",
 		ProjectCommandBuilder: &events.DefaultProjectCommandBuilder{
-			ParserValidator:     &yaml.ParserValidator{},
-			ProjectFinder:       &events.DefaultProjectFinder{},
-			VCSClient:           e2eVCSClient,
-			WorkingDir:          workingDir,
-			WorkingDirLocker:    locker,
-			AllowRepoConfigFlag: "allow-repo-config",
-			AllowRepoConfig:     true,
-			PendingPlanFinder:   &events.DefaultPendingPlanFinder{},
-			CommentBuilder:      commentParser,
+			ParserValidator:   parser,
+			ProjectFinder:     &events.DefaultProjectFinder{},
+			VCSClient:         e2eVCSClient,
+			WorkingDir:        workingDir,
+			WorkingDirLocker:  locker,
+			PendingPlanFinder: &events.DefaultPendingPlanFinder{},
+			CommentBuilder:    commentParser,
+			GlobalCfg:         globalCfg,
 		},
 		DB:                boltdb,
 		PendingPlanFinder: &events.DefaultPendingPlanFinder{},